Fixing examples
The node field nodeTLSSecret is stopping the creation of
the certificates and then the secret does not exist, and the
cluster does not start. I'm figuring out what exactly this feild
is for, and I think it might be for custom cert creation.

Will test the examples more, but I think that they are working now