Scripting to create a GKE Cluster and install operator (#38)
* Scripting to install operator

This PR includes some shell scripts that allow for the creation of an GKE cluster
and pushing the operator container to GCP.
This work will eventually create and install the operators, but I am doing the work
in sections.

* More work on bash scripts

* More work on scripting

Now able to get a basic cluster working, but still have some cleanup to work on.

* Apply operator

- Added functionity to look up the latest image
- Fixed README.md

* Fixing problem with directory resolution

- because we are doing a cd to get kustomize to work, we
are ending up in the wrong directory unless you run it from the
root git directory.  This change fixes that.
- we probably should create a dynamic kustomize files so that
we are not checking in new ones

* Fixing examples

The node field nodeTLSSecret is stopping the creation of
the certificates and then the secret does not exist, and the
cluster does not start. I'm figuring out what exactly this feild
is for, and I think it might be for custom cert creation.

Will test the examples more, but I think that they are working now

* Updating examples

All of them are running now

* Increased cluster size

We need a bigger cluster to run all of the examples at the same time

* fixing name in Makefile

* Fixing comment messages