Updating e2e test to remove annotations
Often annotations are added to pods by other sources. When
testing with gke with network policies enabled the unit tests are
failing because annotations are added by Calico. I am removing
those annotations in this test.

The GKE cluster that I am running the e2e tests has Calico running which
adds annotations to the Pod.  This is causing out e2e test to fail.