Changing e2e testing to use kind
We previously allowed for gke testing during e2e, but
we are going to use kind.  This removes a couple of YAML
stanzas in the test data that kind does not produce.

Fixed path issues and added the kubetest2 binaries.
We should improve this to a toolchain and get osx
support working, but this is a quick fix.