RedHat testing image name override
The statefulset builder now checks if the environment variable
RELATED_IMAGE_COCKROACH is set.  If RELATED_IMAGE_COCKROACH is
set the container image is set to that value instead of the
value set in the CR. RedHat testing sets this evn variable
to test a specific image from a specific container registry.