Using HOSTNAME evn variable
Instead of using the hostname -f command to determine the Pod's
hostname. We are using some bash foo to get the dns name,
the downward api to get the name of the Pod, and the cluster
name to get the name of the serivce.
Various K8s distros and container distros cannot use
hostname -f.