Updating Command and Arguments for CR Pods
This is a cleanup and general change of how we are starting the
CR Pods, and also how we are createing the advertise-host option.
Moreover we are not using a shell command, but are directly running
the coachroach command.  This change is required due to the fact
that the ubi redhat container does not have the hostname command.