Feature Gate Functionality
This PR introduces feature gate functionality to this operator.
We are using code that is currently used within the Kubernetes
codebase.

We now have a command line flag where we can turn on or off
features at runtime and also tag them for stages of release,
for instance Alpha

An example of using the feature flags is running the operator
with the flag `--feature-gates=RunAsNonRoot=false`.  This
flag would shut off the RunAsNonRoot feature gate, and when used
within the code would allow the developer to programmatically
turn off or on new features.