Fixing stamping in container
I turned on stamping and that allows us to use ENV variables
to set to of the labels. I also updated some of the capitalization
in some of the english in the other labels.

The 'version' is now set to the number of commits on the current
branch which will increase.