replace tls.crt and tls.key to our internal node.crt and node.key naming convention
The database requires the files presented to the database for node to node mTLS be called node.crt and node.key.  For some reason the operator has been generating keys named tls.crt and tls.key and then renaming them as they get projected into the statefulset.  This PR brings us in compliance with the naming standard used by the rest of our tooling.