Adding Cockroach DB Version Validator
The operator was originally designed to use the container image tag,
which caused major limitations and problems with Openshift and updates.
This PR implements the operator's capability to determine the version of
the CR binary inside the container, store that information in metadata,
and drives updates off of that metadata.