StatefulSet Changes - CR Cache, Memory, and sts pod management
1. Dynamically set cache and max sql memory if the user does
not set a CRD value
2. sts is set to Parallel pod management

We are making changes to use `exec` and bash directly in order to
start the Pods and set the cache values dynamically.  We are also
adding environment variables in to the Pods that set the GOPROCSMAX
and also MEMORY_LIMIT_MIB which is used to set CR cachee and
max sql memory values.