Creating RBAC objects for Openshift E2E
This PR creates the RBAC bindings that are required to run e2e testing
with openshift. We are creating the role binding and role that
allows the operator database service account to a Pod as any UID.

When the e2e test starts the role is created, and when the test
is over the role is destroyed.  This code stop e2e testing
to occur in parallel. But we have other problems with parallel testing
as well.