Do not override `CHANNEL` and `IS_DEFAULT_CHANNEL` (#627)
Previously, we hardcoded the `CHANNEL` and `IS_DEFAULT_CHANNEL` variables
in Makefile for the `release/gen-files` target in order to reduce the
possibility of publishing a release to a wrong channel.

Now, that we have these steps documented and automated, we can remove
the hardcoded variables. By default, they will be set to `CHANNEL=beta`
and `IS_DEFAULT_CHANNEL=0`. The release docs will explicitly override
these variables.