Update env.Env to be universal
For most things, we're using `env.Env` as the test environment, which
performs a `paths.MaybeSetPath` using kubetest2-kind as the lookup
variable. This will check the _PATH_ to see if the parent directory is
included and add it if it's missing.

**The problem:**

The api tests (for example) use the standard envtest provided by
controller-runtime to create an environment. As we beef up the env in
this project, it'd be nice to leverage the same env for all tests.

Also, it seems that all e2e tests and the api ones require that a few
executables be available, more specifically things like kubectl,
kubeapi-server, and kind. Rather than looking for `kubetest2-kind` and
doing a `MaybeSetEnv` or individually setting `TEST_KUBECTL`, etc. I
think we could leverage `KUBEBUILDER_ASSETS` and `PATH` automatically.

**The solution:**

Drop the _testutil/paths_ package and add `env.ExpandPath` as well as
`env.PrependPath`. These are then used to set `KUBEBUILDER_ASSETS` to the
_hack/bin_ directory in the test workspace (also works without bazel)
as well as prepend that path to PATH.

This path is defined by the `binaries` flag, which can be
overridden if necessary using the `args` key in a `go_test` target.