Add kuberecord to enable VCR for K8s API calls (#676)
Adding `kuberecord` which is a package that leverages the
[go-vcr](https://github.com/dnaeon/go-vcr) to record interactions with a
K8s API server.

This is useful to avoid having to run every command over and over again
when running tests. In theory, unless the API version changes, we
typically don't expect the interaction to be any different.

Kuberecord works by wrapping the incoming rest.Config to handle loading
prerecorded interactions, scrubbing sensitive/flakey headers, and
storing results to disk (`./testdata/<name>.vcr` by default).

See _record_test.go_ in this PR for an example of how to use it.