Add defaulting and validating webhooks
There are a number of places where we perform some validation of the
CrdbClusterSpec. While some of this is more complicated, most of the
checks are for setting some defaults or checking that a certain value is
within some bounds.

Rather than doing this in our actors/resources, I figured we could
leverage the admission controller by providing a couple of webhooks to
handle setting defaults and validing fields on the way in. This
centralizes the basic validation logic and ensures we don't need to have
checks for these invariants throughout the code base.

This change adds a very basic implementation of webhook.Defaulter and
webhook.Validator and sets it up to work with the manager at startup.
For now, there's no special validation logic in place. Over time we can
build that up.

All communication between the K8s API and the webhook service must use
TLS. By default, a self-signed certificate will be generated and made
available when the controller starts up. Alternatively, an operator
could set their own certs by creating the
`cockroach-operator-webhook-tls` secret, setting `tls.key` and `tls.crt`
with their respective PEM-encoded values.