Updating markers for cluster types (#679)
Fixes #668

There are a couple of places where we've got `+required` markers on
struct fields, but have `omitempty` in the JSON tag. Similarly we've got
a few mismatched `+optional` and `(Required)` text in the comments.

This patch addresses those and standardizes on the following format:

```
[(Optional)] <description>
[Default: <val>]
[other markers]
[+optional|+required]
```

I ran `make dev/generate` after making these changes to ensure the
generated code was up to date.