Add make commands for running locally
While testing out some validation locally, I noticed that our markers
weren't correct for the VolumeClaim type. It expects, _**either**_ the
spec or the source to be set, but not both. This means the both need to
be marked optional.

In order to make this process easier to test, I've added `make dev/run`
and `make dev/run-shell` targets. These run a local kind cluster (with
the operator pre-installed) and a pre-configured shell respectively.

I've updated DEVELOPER.md with details about these targets.