Use OpenAPI validation for node count
We currently check for the node count to be >=3 in the decommissioner.
There's actually a `TODO` comment in there about adding a validation
hook to avoid this. However, since the admission controller will always
run the OpenAPI validation when resources are submitted to the API
server, we don't actually need a hook for this.

I've added a minimum=3 marker to the spec, which will be enforced
automatically, negating the need for this check.