Removing some unused or unnecessary hack scripts
We have a bunch of targets in //hack that check if they're running in
bazel and either warn the user to install bazel or subsequently calling
itself via bazel run/test.

If we remove the scripts (where possible), we can avoid the need to have
these checks, as well as configuring an explicit target in the BUILD
files.