Drop Nodes() check and loop through builders in deploy action (#721)
There's an unnecessary check for the node counts here in the deploy
action. This is handled automatically by the admission controller since
we've got a validation on the minimum node count (3).

While I was here, I also updated the action to loop through all the
builders (in order) to avoid duplicating all of the reconciler and error
checking code.