Refactor packaging scripts
We use two main scripts for packaging up the bundle;
update-pkg-manifest.sh and opm-build-bundle.sh. The former generates
the CSV and CRD files in deploy/.../<version>, while the later generates
a bundle and updates the latest release info.

I found these scripts, as they were written, a little hard to grok. So I
updated them to use smaller, well-named functions that state what the
goal is and hopefully make the process a little easier to follow for
others as well.