Leverage customize for generating manifests (#748)
We use two different ways to generate manifests right now. In
development, kustomize bases from config are used. However in our
production documentation, we ask users to apply a manifest in
_manifests/operator.yaml_ which is generated with a template.

There were a number of issues I found here. Particularly, the RBAC
permissions were different in the template and the generated rbac base.
This was mainly due to the fact there were a lot of bindings for the
cluster role defined in the template, but not as `//+kubebuilder:rbac`
comments in the controller.

I've created a number of kustomizes bases here. At a high level, the
three that are important for this PR are; default, manifests, and
operator.

**Default**

This base contains everything you'd need to run the operator. That
includes CRDs, deployments, RBAC roles/bindings, and the webhooks. This
is used locally when running `make dev/up` for example.

**Manifests**

This base contains the generated CSV as well as patches for OpenShift.
It includes everything from default as well as samples, scorecard and a
patch to setup all of the env vars (e.g.
`RELATED_IMAGE_COCKROACH_v20_1_4`) we need for deployment.

**Operator**

This is the base used to generate _install/operator.yaml_. (The root
manifests directory has been renamed here to install). Similarly,
install/crds.yaml is generated from the config/crds base now.

We'll need to update our docs to reflect this, but this is same atm
since we recently updates the docs to point to a tagged release rather
than master directly.