Adding CHANGELOG.md for tracking changes between releases
The title says it all really. The format follows a known standard, and
includes links to the diffs for every released version.

I've also updated hack/release to automatically conver the previous
`[Unreleased]` section to the new release and add a new unlreleased
heading to point to the diff between the latest release and master.

Finally, I went through the changes from v2.1.0 to now and backfilled
the details here.

If we've got a previous beta, as we do right now, we'll need to copy the
details (Added/Changed/etc) from those to the new release before pushing
the new release branch.

New PRs will have a checklist that will remind us to update the
changelog regularly.