Fix issue with make dev/up (#754)
During the process of kustomizing everything and updating the release
scripts there was an issue introduced that broke `make dev/up`.

The issue is that while the `:latest` is defined in
_config/manager/deployment.yaml_, there's a patch for the image that
references the latest release explicity (`v2.2.0-beta.2`).

I think we want to use a pinned version for everything except
development, so rather than remove the patch, I've updated _dev.sh_ to
pass in the version to be replaced.

One issue I ran into was not being able to use the stamp values as
replacements. So instead I used `--define APP_VERSION=$(APP_VERSION)`
which will make the variable available for make substitution.