Update CertificatesV1Beta1 to CertficatesV1
Kubernetes 1.22 has dropped the v1beta version of the Certificates API.
See
<https://kubernetes.io/blog/2021/07/14/upcoming-changes-in-kubernetes-1-22/#api-changes>
for more information.

Fortunately for us, we're not actually using the code we have here that
references the v1beta1 API. So I've removed it.

In preparation for OpenShift 4.9 (which uses k8s 1.22) I've verified
that were' not using any of the removed APIs from the list above.