Address issues with running on OpenShift
We have a couple of issues that prevent the operator from running
properly on OpenShift.

**Issue 1: Webhooks**

The first has to do with webhook configuration. On startup, we generate
a CA and associates TLS certificates for the webhook-service. The CA is
patched into the webhook config's CABundle and the certs are placed on
disk for the webhook service to use.

The trouble is that openShift doesn't name the webhook configurations
the same (for the better IMO). So at startup we can't find the hook and
bail on startup (CrashLoopBackoff).

OpenShift will setup the service and patch the CA bundle automatically,
so I've added the `-skip-webhook-config` flag and ensured that it's set
in the manifests for OpenShift by default.

**Issue 2: Jobs**

When we deploy a CrdbCluster, we queue a version checker job. There are
currently permission issues with this since we don't have update
permissions on crdbclusters/finalizers. More details can be found here:
https://kubernetes.io/docs/reference/access-authn-authz/admission-controllers/#ownerreferencespermissionenforcement

I've added the appropriate permissions here.