Allow operator to work for multiple namespaces (#815)
Recently #812 was merged, which updated the operator to allow it to
handle multiple CRs in the same namespace. This PR builds on that by
allowing the operator to work across multiple namespaces.

By default, the operator will be installed in the
`cockroach-operator-system` namespace and be set up to listen for
CrdbCluster resources in all namespaces. This is a departure from our
current setup of being in the default namespace and being limited to
that namespace alone.

Rather than enforcing a cluster-wide operator for all use cases, I've
made it configurable. Using the `WATCH_NAMESPACE` env var, you can
control the install mode of the operator.

* _empty_ or not defined (default) - watch all namespaces
* `myns` - only watch for resources in `myns` namespace
* `ns1,ns2` - watch for changes in `ns1` and `ns2` only

When installing the operator from OperatorHub within an OpenShift
cluster, there will be an option to select either _all_ namespaces, or a
_single_ one.