Update code that generates the certified bundles
There's been a change to the way OpenShift bundles are certified which
changes our release process.

I've updated the code here to drop the bundles into the following
directories, which represent the certified bundle and the marketplace
bundle respectively.

* `bundle/<version>/cockroachdb-certified`
* `bundle/<version>/cockroachdb-certified-rhmp`

After we cut a release (via `make release/new`), we'll need to perform
the following steps:

* Publish docker image (no changes)
* Tag the release (no changes)
* Upstream PR with the certified bundle (e.g.
  https://github.com/redhat-openshift-ecosystem/certified-operators/pull/81)
* Upstream PR with the marketplace bundle (e.g.
  https://github.com/redhat-openshift-ecosystem/redhat-marketplace-operators/pull/5)