Give nightly tests a bit longer to be ready
Our nightly tests have been failing. At first, we thought it might just
be flakiness, but the issue appears to be consistent.

After running the tests from my local machine, I was able to reproduce
the issue and found that by adding a pause before saying K8s is ready, I
can reliably get the tests to pass.

Clearly, adding a sleep call is not ideal/desireable. I'm hoping to get
this in so we can stop the bleeding so to speak and then figure out what
the actual issue is.