Don't store bundle here (#840)
With the updated release process for OperatorHub we no longer need to
keep the bundle around in this repo, since these will be in their
respective upstream repos.

When generating the bundles, we also need to get the SHA of the image
since certified bundles no longer allow tags (since these can be changed
after the fact). In the current process, the image hasn't been released
yet, so we're unable to generate a valid bundle at this point anyways.

I've updated the code here to skip the bundle when running `make
release/new`. The updated release process will now be:

* `make release/new VERSION=x.y.z` (+ PR->Review->Merge)
* Publish docker image
* Tag the release
* Run `make release/generate-bundle`
* Use dirs in ./bundle (now gitignored) to create upstream PRs for the
  certified and certified-rhmp releases.