Update Release Process for OpenShift (#843)
While releasing v2.5.0 I ran into a number of issues, particularly with
the marketplace bundle. Marketplace bundles are now required to
explicitly list any external docker images (pinned by SHA) in the CSV.

This PR represents all of the changes that were made in order to get a
working bundle into the Marketplace. Running `make
release/generate-bundle` will create both the certified and the
marketplace bundles in _./bundle_ which can be copied into the
corresponding locations in the certification repos when publishing a new
release.

Notable changes:

* Add `relatedImages` to the CSV spec with all of the crdb images listed
* Updated operator-sdk to v1.15
* Update crdb versions and set default to v21.2.3