Ensuring Openshift binaries for both OSX and Linux (#847)
* Ensuring openshift binaries for both OSX and Linux

When running make release/generate-bundle on a linux machine it didn't
go as expected. Some of the SHAs for the openshift binaries were wrong
due to the fact that pinning to `stable-<major>.<minor>` allows for
changes in the underlying versions (and therefore the SHAs).

This PR pins openshift binaries to v4.9.17 and adds some structure to
allow making updates in the future easier.

I've also updated the `sed` commands in hack/bundle.sh to avoid using
`-i ''` which only seems to work on OSX. Simply redirecting the output
to a new file and then moving it to the original path will work
cross-platform with or without GNU sed.