Add RELEASE_IMAGE_COCKROACH_<version> to manager env (#848)
When supplying `cockroachDBVersion` in the spec rather than a specific
image, we check for the image in the manager's env. However, at the
moment that env doesn't include the related images.

This PR addresses that by adding them to the generated deployment
manifest, which in turn is used to create the operator.yaml file in
_/install_.

I've verified that everything still works as expected when generating
the bundle by running `make release/generate-bundle` and ensuring the
image values are what we expect.