Replace kind with k3d
When running some tests I was able to reduce the time they take to run
quite a bit by using k3d rather than kind.

In smaller tests, this was significant, for example the smoketests ran
in less than half the time. On larger tests, it shaves a couple of
minutes off (on the ~10 I ran anyways). Pressumably this is just the
difference in the time it takes to start/stop k3d vs kind.

To make this all work, I had to create a custom kubetest2 deployer for
k3d. It's in _e2e/kubetest2-k3d_ and largely is a copy of the kind
deployer updated to use k3d instead.