Validate pvc.volumeMode to circumvent potential reconcilation loop
This is an attempt to fix the issue reported
in https://github.com/cockroachdb/cockroach-operator/issues/850
and https://github.com/cockroachdb/cockroach-operator/issues/845.

Currently, there seems to be a bug in
[k8s-objectmaker](https://github.com/banzaicloud/k8s-objectmatcher/pull/49),
which causes the operator to think that the sts
has changed since creation and get stuck in the reconcilation loop.

To mitigate, we check whether `volumeMode` is set,
and set correctly.