Enable support for leader election.
As mentioned in #863, there is no longer a default value for
`LeaderElectionID` on the manager options. The means that when
`--enable-leader-election` is supplied, the operator fails to start.

I've supplied a default value for the flag here, since most cases won't
need to change this. However, if you intend on running multiple versions
of the operator, having the CLI flag makes it easy to avoid name
collisions.

Fixes #863