Increase timeout for cluster creation in E2E tests (#870)
Our e2e tests have been failing. From looking at the logs, there are a
number of things that could be the cause.

The first, is that we're seeing cluster creation timeouts. To mitigate
this I've added a shared variable we can use to set the timeout as
necessary. For now, I've bumped it from 500s to 10m (600s).

The second issues is that it seems our openshift kubetest2 thingy
doesn't clean up properly once something fails. This means we need to go
into the Google console and delete some DNS record sets, NAT tables, and
load balancers. This should be updated, but isn't addressed here.