Bump e2e cluster creation timeout to 20m (#871)
We continue to experience timeouts when running GKE nightlies. The
errors are timeouts, particularly in the upgrade e2e test. I don't see
anything obvious in the code, nor any meaningful changes from when they
worked to fix. So I've doubled the timeout from 10m to 20m.

If this works, as in, it provides the necessary amount of time, we can
tighten it over time.