Fix broken CI (#875)
We have a couple of issues that are preventing CI runs from passing.
These have been addressed here. I've also removed setup-bazelisk from
the workflows, since this is installed by default in the VM that is used
to run GH actions. I've also bumped the actions we use to their
respective latest versions.

**zlib 1.2.11 is no more**

A new version of zlib (1.2.12) was released recently (March 27). This
release is meant to replace v1.2.11 and as such, the old version was
yanked. While we don't directly depend on this, our bazel WORKSPACE
brings in protobuf, which in turn includes zlib. Bumping this from 3.9.0
to 3.20.0, resolved the issue.

**RedHat API is different**

Previously when we fetched the crdb image versions from the RedHat Catalog
API, we were able to limit the returned results to just the image digest
and the repo tags. This no longer works as it appears that filtering is
now limited to a single level (can filter for `data.repositories`, but
not `data.repositories.tags.name` as we used to).