Increase TerminationGracePeriod to 5m
Currently we have the termination grace period set to 1m. This seems a
bit low since we need to do a number of things (e.g. drain connections)
before properly shutting down.

I've bumped this to 5m to give the nodes a little more time. This
represents an upperbound of the time, so should have no speed/perf
penalty in most cases.