Run only upgrades e2e test for nightlies
Our nightlies have been (sometimes) timing out for a while now.
Currently we run create, update, and decommission e2e tests for the
nightly CI runs for GKE and OpenShift.

This PR reduces that list to just the upgrades test. This involves
creating a cluster, updating the version, and decommissioning it, which
should be sufficient for us. We still run each of the e2e tests in CI,
so this doesn't represent a reduction in overall test coverage, just
limits what's being tested in the nightly run.