Allow option to automount service account token (#898)
In #107 auto-mounting of the default service account token was disabled.
Presumably this was because it wasn't necessary in order to run
Cockroach.

We've had a number of users request the ability to re-enable this as
it's the default behaviour in Kubernetes and is often used to grant
access to other systems through IAM grants (e.g. backups in S3).

This PR adds an `AutomountServiceAccountToken` field to the cluster
spec. When set, the service account token will be mounted. By default
this value is `false` for backward compatibility reasons.