Update Openshift Release Process (#964)
* Install preflight binary on Linux machines

Adding preflight to the list of OpenShift binaries installed. For now,
this is only available on Linux machines and won't be added to anyone
using Darwin.

* Publish operator and bundle images for RedHat/OpenShift

Updating the build/release scripts to handle publishing the bundle
images and running preflight on them.

The publish-openshift script downloads the tagged release from
DockerHub, retags it correctly for the RedHat scanning repo, pushes it,
and finally runs the preflight check on it.

The openshift-bundle script largely does the same, only it publishes the
bundle image rather than the operator image. The same script is used for
both the certified and marketplace bundles. Passing `MARKETPLACE=1` will
use the marketplace bundle values rather than the certified ones
(default).

* Update OpenShift binaries to 4.10.18

Updated all of the outdated OpenShift binaries. This also includes a fix
to an issue with had with opm. We were copying the tar file into the bin
folder and marking it as executable. This naturally, caused issues when
we tried to run it.

* Add support for make test/preflight-<target>

Adding targets for running preflight tests. These only work on Linux
machines since preflight is currently not available on other platforms.

The preflight tasks all assume you've got a running OpenShift cluster
and that GCP_PROJECT and KUBECONFIG are set accordingly in the env.

* Build, Publish and automate release process for openshift

* Fix the review comments

---------

Co-authored-by: David Muto (pseudomuto) <david.muto@gmail.com>